Moved purchases data to entitlements context and general listeners for it to App.js. Updated PurchaseScreen and LoginScreen to reflect new context.
