+++ constants/EntitlementsContext.js
@@ -1,3 +1,3 @@
 import { createContext } from "react";
 
-export const Entitlements = createContext({ storage1: false });
+export const Entitlements = createContext({ isAnon: true, revCatId: null, storage1: false });
