+++ screens/PurchaseScreen.js
@@ -1,6 +1,6 @@
 import { useContext, useEffect, useState } from "react";
 import { SafeAreaView, View, StatusBar, FlatList, Alert, Text } from "react-native";
-import Purchases, { LOG_LEVEL } from "react-native-purchases";
+import Purchases from "react-native-purchases";
 import ButtonBar from '../components/ButtonBar';
 import Icons from "../constants/Icons";
 import Colors from "../constants/Colors";
@@ -18,23 +18,12 @@ export default function PurchaseScreen ({ route, navigation }) {
     const [packages, setPackages] = useState([])
     const [isPurchsing, setIsPurchasing] = useState(false);
     const [isAnonymous, setIsAnonymous] = useState(true);
-    const [revCatId, setRevCatId] = useState(null);
-    const [subscriptionActive, setSubscriptionActive] = useState(false);
 
     useEffect(() => {
-        // Purchases.setLogLevel(LOG_LEVEL.VERBOSE);
         const getPackages = async () => {
-            // if (Platform.OS === 'ios') {
-            //     Purchases.configure({apiKey: "appl_NIMzKbuELZwYrRadlznGbomLWLN"});
-            // } else if (Platform.OS === 'android') {
-            //     Purchases.configure({apiKey: "goog_vCtRNkrJEMHsuLzlXyAtVaRsWjq"});
-            // }
             try {
                 const offerings = await Purchases.getOfferings();
-                // offerings !== undefined ? console.log(offerings) : console.log("Offerings are undefined")
                 if (offerings.current !== null && offerings.current.availablePackages.length !== 0) {
-                    // console.log("Available Packages: ")
-                    // console.log(offerings.current.availablePackages)
                     setPackages(offerings.current.availablePackages);
                 }
             } catch (e) {
@@ -45,27 +34,24 @@ export default function PurchaseScreen ({ route, navigation }) {
           getPackages();
     }, []);
 
-    const getUserDetails = async () => {
-        setIsAnonymous(await Purchases.isAnonymous());
-        setRevCatId(await Purchases.getAppUserID());
+    // const getUserDetails = async () => {
+    //     setIsAnonymous(await Purchases.isAnonymous());
     
-        const customerInfo = await Purchases.getCustomerInfo();
-        if (typeof customerInfo.entitlements.active[Strings.util.entitlements.storage1] !== 'undefined') {
-            let ent = {...entitlements};
-            ent.storage1 = true;
-            setEntitlements(ent);
-            setSubscriptionActive(true);
-        }
-        // setSubscriptionActive(typeof customerInfo.entitlements.active[Strings.util.entitlements.storage1] !== 'undefined');
-    };
+    //     const customerInfo = await Purchases.getCustomerInfo();
+    //     if (typeof customerInfo.entitlements.active[Strings.util.entitlements.storage1] !== 'undefined') {
+    //         let ent = {...entitlements};
+    //         ent.storage1 = true;
+    //         setEntitlements(ent);
+    //     }
+    // };
     
-    useEffect(() => {
-        // Get user details when component first mounts
-        getUserDetails();
-    }, []);
+    // useEffect(() => {
+    //     // Get user details when component first mounts
+    //     getUserDetails();
+    // }, []);
 
     useEffect(() => {
-        if (subscriptionActive) {
+        if (entitlements.storage1) {
             if (isAnonymous || !user.uid) {
                 let obj = {...settingsObj}
                 if (obj.login === Strings.util.logins[0]) {
@@ -77,15 +63,15 @@ export default function PurchaseScreen ({ route, navigation }) {
                 navigation.push(Strings.util.routes.home)
             }
         }
-    }, [subscriptionActive]);
+    }, [entitlements.storage1]);
     
-    useEffect(() => {
-        // Subscribe to purchaser updates
-        Purchases.addCustomerInfoUpdateListener(getUserDetails);
-        return () => {
-          Purchases.removeCustomerInfoUpdateListener(getUserDetails);
-        };
-    });
+    // useEffect(() => {
+    //     // Subscribe to purchaser updates
+    //     Purchases.addCustomerInfoUpdateListener(getUserDetails);
+    //     return () => {
+    //       Purchases.removeCustomerInfoUpdateListener(getUserDetails);
+    //     };
+    // });
 
     let cancelBtn = {
         title: Strings.English.buttons.cancel,
@@ -134,8 +120,8 @@ export default function PurchaseScreen ({ route, navigation }) {
                             }
                         } catch (e) {
                             if (!e.userCancelled) {
-                                // Alert.alert(Strings[language].headers.errorAlert, e.message)
-                                Alert.alert("Purchase Item Error: ", e)
+                                Alert.alert(Strings[language].headers.errorAlert, e.message)
+                                // Alert.alert("Purchase Item Error: ", e)
                             }
                         } finally {
                             setIsPurchasing(false);
@@ -146,9 +132,6 @@ export default function PurchaseScreen ({ route, navigation }) {
                 />}
                 keyExtractor={(item) => item.identifier}
             />}
-            {/* <Text>{revCatId}</Text>
-            <Text>{subscriptionActive ? 'Active' : 'Not Active'}</Text>
-            <Text>{isAnonymous ? 'Anonymous' : 'Identified'}</Text> */}
         </View>
         <ButtonBar buttons={[cancelBtn, restoreBtn]} />
     </SafeAreaView>
